+++ graphbrain/cognition/agents/actors.py
@@ -2,10 +2,10 @@ import logging
 from collections import Counter
 
 from graphbrain.cognition.agent import Agent
-from graphbrain.meaning.concepts import has_proper_concept
-from graphbrain.meaning.concepts import strip_concept
-from graphbrain.meaning.corefs import main_coref
-from graphbrain.meaning.lemmas import deep_lemma
+from graphbrain.utils.concepts import has_proper_concept
+from graphbrain.utils.concepts import strip_concept
+from graphbrain.utils.corefs import main_coref
+from graphbrain.utils.lemmas import deep_lemma
 from graphbrain.op import create_op