+++ graphbrain/cognition/agents/corefs_dets.py
@@ -1,8 +1,8 @@
 import logging
 
 from graphbrain.cognition.agent import Agent
-from graphbrain.meaning.concepts import has_proper_concept
-from graphbrain.meaning.corefs import make_corefs_ops
+from graphbrain.utils.concepts import has_proper_concept
+from graphbrain.utils.corefs import make_corefs_ops
 
 
 class CorefsDets(Agent):