+++ graphbrain/cognition/agents/corefs_onto.py
@@ -1,8 +1,8 @@
 import logging
 
 from graphbrain.cognition.agent import Agent
-from graphbrain.meaning.corefs import make_corefs_ops
-from graphbrain.meaning.ontology import subtypes
+from graphbrain.utils.corefs import make_corefs_ops
+from graphbrain.utils.ontology import subtypes
 
 
 class CorefsOnto(Agent):