+++ graphbrain/notebook.py
@@ -4,6 +4,7 @@ from IPython.core.display import display
 from IPython.core.display import HTML
 
 from graphbrain import hedge
+from graphbrain.extractors.case_generators import InteractiveCaseGenerator
 
 
 TYPE_COLORS = {'C': '#268bd2',
@@ -258,3 +259,8 @@ def blocks(edge, subtypes=False, argroles=True, namespaces=False):
     html = '<div style="background-color:#fcfcfc; padding:50px">{}'\
            '</div>'.format(html)
     display(HTML(html))
+
+
+def interactive_case_generator(tests, extractor, case=None, keywords=None):
+    InteractiveCaseGenerator(tests, extractor).run(case=case,
+                                                   keywords=keywords)