+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Sequences keep track of their size, no need to specify position when adding edge to sequence.
 - Improved wikipedia agent.
 - Improved coreference resolution.
+- LevelDB included by default again.
 
 ### Removed